rename src to screen_translate See: http://as.ynchrono.us/2007/12/filesystem-structure-of-python-project_21.html
