+++ screen_translate/Capture.py
